Fixed some warnings, continued working on 64bit conversion..
Committing changes!

Made a number of changes to MAMClient that enabled successful building on both
64bit and 32bit machines. In addition to this, the work performed also enables
successful login of large and small usernames, corrected some memory issues,
and resolved some code completeness issues.

+++ cpp/MAMClient/CryptoStuff.h
@@ -24,8 +24,8 @@ public:
     CryptoStuff( void );
     ~CryptoStuff( void );
 	
-    void incoming( char *packet, int size );
-    void outgoing( char *packet, int size );
+    void incoming( char *packet, size_t size );
+    void outgoing( char *packet, size_t size );
 	
     void generate_keys( CPacket packet );
     void use_new_keys( void ) {	this->m_useNewKeys = true; }
